requirements.txt
asgiref==3.9.0
dj-config-url==0.1.1
dj-database-url==3.0.1
Django==5.2.4
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.9.0
